requirements.txt
streamlit==1.50.0
tensorflow==2.12.0
numpy==1.25.2
opencv-python-headless==4.8.0.74
mediapipe==0.10.3
gTTS==2.3.1
playsound3==3.0.0
huggingface-hub==0.16.4
pillow==10.0.0